Simplify Emotes and emoteSet, Add emote extractor
