+++ TwitchLib.Client.Models/Emote.cs
@@ -1,5 +1,8 @@
 ﻿namespace TwitchLib.Client.Models
 {
+    /// <summary>
+    /// Object representing an emote in an EmoteSet in a chat message.
+    /// </summary>
     public class Emote
     {
         /// <summary>Twitch-assigned emote Id.</summary>
